Remove search feature and add embed page
